requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
decouple==0.0.7
Django==4.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.7
multidict==6.0.4
packaging==23.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
twilio==8.10.3
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
zope.interface==6.1
